Add --useTitles options
--useTitles creates the local directory structure based on the photo
and set titles instead of the Flickr ID's, which makes it easier to
find a specific photo in the backup directory.

Furthermore, this commit modifies the --limitToSet option so that it
always matches set titles.  If --useTitles is enabled, then the
sanitized title is also matched.  Otherwise, the Flickr set ID is also
matched.